Fix a few more code style issues
